refactor signup form to a component
